refactor: 엔티티 이름 패키지 변경
- ConfirmedSchedule -> ConfirmedMeeting
- MeetingController 내부로 로직 이동
- ConfirmScheduleService -> MeetingConfirmService
