+++ backend/src/main/java/kr/momo/service/meeting/dto/MeetingConfirmRequest.java
@@ -1,9 +1,9 @@
-package kr.momo.service.schedule.dto;
+package kr.momo.service.meeting.dto;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDateTime;
 
-public record ScheduleConfirmRequest(
+public record MeetingConfirmRequest(
         @NotNull LocalDateTime startDateTime,
         @NotNull LocalDateTime endDateTime
 ) {
